Revert donationTier primitive type UShort to Int
Unsigned types are still experimental in kotlin and do require a lot of additional changes when using mixed types

+++ app-domain/src/main/kotlin/co/anitrend/domain/user/entity/contract/UserStatus.kt
@@ -29,7 +29,7 @@ package co.anitrend.domain.user.entity.contract
  */
 data class UserStatus(
     val donationBadge: CharSequence?,
-    val donationTier: UShort?,
+    val donationTier: Int?,
     val isFollowing: Boolean?,
     val isFollower: Boolean?,
     val isBlocked: Boolean?,
